Remove debug messages from TeenageAshe, "new" Orbwalker
It turns out it wasn't my fault TeenageAshe wasn't working.
League.Common.Menu.SetValue<T>() has no contigency plan in case it tries
to save on a corrupted file. It doesn't even catch the exception that is
thrown. All I had to do was to delete the caches menu Files.

As for the Orbwalker, it's a dll, and it's called TOrbwalker. I
literally copied LeagueSharp.Common.Orbwalking and added the option for
the behaviours to be toggable (except for Combo).